+++ CHANGELOG.md
@@ -2,6 +2,28 @@
 All notable changes to this project will be documented in this file. This project adheres to
 [Semantic Versioning](http://semver.org/) and [this changelog format](http://keepachangelog.com/).
 
+## [1.2.0] - 2019-06-20
+
+### Added
+- [#360](https://github.com/cloudcreativity/laravel-json-api/issues/360)
+Allow soft delete attribute path to use dot notation.
+- Added `domain` method to API fluent routing methods.
+- [#337](https://github.com/cloudcreativity/laravel-json-api/issues/337)
+Can now apply global scopes to JSON API resources via [adapter scopes.](./docs/basics/adapters.md#scopes)
+
+### Fixed
+- [#370](https://github.com/cloudcreativity/laravel-json-api/pull/370)
+Fix wrong validation error title when creating a custom validator.
+- [#369](https://github.com/cloudcreativity/laravel-json-api/issues/369)
+Fix using an alternative decoding type for update (`PATCH`) requests.
+- [#362](https://github.com/cloudcreativity/laravel-json-api/issues/362)
+Fix fatal error in route class caused by polymorphic entity types.
+- [#358](https://github.com/cloudcreativity/laravel-json-api/issues/358)
+Queue listener could trigger a `ModelNotFoundException` when deserializing a job that had deleted a
+model during its `handle()` method.
+- [#347](https://github.com/cloudcreativity/laravel-json-api/issues/347)
+Update `zend-diactoros` dependency.
+
 ## [1.1.0] - 2019-04-12
 
 ### Added